Add Authorization on UserTrack endpoints

+++ src/main/java/org/ppke/itk/listen/controller/UserTrackController.java
@@ -8,6 +8,9 @@ import org.ppke.itk.listen.domain.user.repository.UserRepository;
 import org.ppke.itk.listen.domain.usertrack.data.UserTrack;
 import org.ppke.itk.listen.domain.usertrack.repository.UserTrackRepository;
 import org.ppke.itk.listen.model.NewUserTrack;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -51,46 +54,79 @@ public class UserTrackController {
     public UserTrack saveNewTrack(@RequestBody NewUserTrack userTrack) {
         log.info("Calling POST /user-tracks endpoint.");
 
-        User ownerUser = userRepository.findById(userTrack.getOwnerUserId())
-                .orElseThrow(() -> new NoSuchElementException("Could not find " + User.class.getSimpleName() + " with the id: " + userTrack.getOwnerUserId()));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentPrincipalName = authentication.getName();
+        log.info("User authenticated");
+
+        User authenticatedUser = userRepository.findByName(currentPrincipalName)
+                .orElseThrow(() -> new NoSuchElementException("Could not find " + User.class.getSimpleName() + " with the name: " + currentPrincipalName));
+        log.info("User entity found");
 
         final UserTrack newTrack = UserTrack.builder()
             .name(userTrack.getName())
             .audioUrl(userTrack.getAudioUrl())
             .pictureUrl(userTrack.getPictureUrl())
-            .ownerUser(ownerUser)
+            .ownerUser(authenticatedUser)
             .build();
+        log.info("New UserTrack created");
 
         return userTrackRepository.save(newTrack);
-
-        // TODO: Only allow track upload for authenticated user for their ids
     }
 
     @PutMapping(value ="/{id}")
     public UserTrack updateTrack(@RequestBody NewUserTrack userTrack, @PathVariable("id") Long id) {
         log.info("Calling PUT /user-tracks/{id} endpoint.");
 
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentPrincipalName = authentication.getName();
+        boolean hasAdminRole = authentication.getAuthorities().stream()
+          .anyMatch(r -> r.getAuthority().equals("ROLE_ADMIN"));
+
+        log.info("User authenticated");
+
+        User authenticatedUser = userRepository.findByName(currentPrincipalName)
+                .orElseThrow(() -> new NoSuchElementException("Could not find " + User.class.getSimpleName() + " with the name: " + currentPrincipalName));
+        log.info("User entity found");
+
         UserTrack trackToUpdate = userTrackRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("Could not find " + UserTrack.class.getSimpleName() + " with the id: " + id));
+        log.info("UserTrack is found");
+
+        if(authenticatedUser != trackToUpdate.getOwnerUser() & !hasAdminRole) {
+            throw new AccessDeniedException("Authenticated user is not authorized to edit UserTrack");
+        }
 
         trackToUpdate.setName(userTrack.getName());
         trackToUpdate.setAudioUrl(userTrack.getAudioUrl());
         trackToUpdate.setPictureUrl(userTrack.getPictureUrl());
+        log.info("UserTrack updated");
 
         return userTrackRepository.save(trackToUpdate);
-
-        // TODO: Only allow track update for authenticated user for their ids
     }
 
     @DeleteMapping(value ="/{id}")
     public void deleteTrack(@PathVariable("id") Long id) {
         log.info("Calling DELETE /user-tracks/{id} endpoint.");
 
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentPrincipalName = authentication.getName();
+        boolean hasAdminRole = authentication.getAuthorities().stream()
+          .anyMatch(r -> r.getAuthority().equals("ROLE_ADMIN"));
+        
+        log.info("User authenticated");
+
         UserTrack trackToDelete = userTrackRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("Could not find " + UserTrack.class.getSimpleName() + " with the id: " + id));
+        log.info("UserTrack is found");
+
+        User authenticatedUser = userRepository.findByName(currentPrincipalName)
+                .orElseThrow(() -> new NoSuchElementException("Could not find " + User.class.getSimpleName() + " with the name: " + currentPrincipalName));
+        log.info("User entity found");
+
+        if(authenticatedUser != trackToDelete.getOwnerUser() & !hasAdminRole) {
+            throw new AccessDeniedException("Authenticated user is not authorized to delete UserTrack");
+        }
 
         userTrackRepository.delete(trackToDelete);
-        
-        // TODO: Only allow track delete for authenticated user for their ids
     }
 }
